+++ ec2_metadata.py
@@ -6,7 +6,7 @@ from cached_property import cached_property
 
 __author__ = 'Adam Johnson'
 __email__ = 'me@adamj.eu'
-__version__ = '1.0.0'
+__version__ = '1.1.0'
 
 __all__ = ('ec2_metadata',)