+++ HISTORY.rst
@@ -8,7 +8,7 @@ Pending Release
 
 .. Insert new release notes below this line
 
-* Add security groups attribute.
+* Add ``security_groups`` attribute.
 
 1.0.0 (2017-06-16)
 ------------------