+++ HISTORY.rst
@@ -8,6 +8,9 @@ Pending Release
 
 .. Insert new release notes below this line
 
+1.1.0 (2017-08-07)
+------------------
+
 * Add ``security_groups`` and ``user_data`` attributes.
 
 1.0.0 (2017-06-16)